Fix module not found by moving app.py to /PFT
When running the streamlit version of the app, there was a module not
found error when trying to import any modules from PFT. This was because
app.py was located in PFT/PFT. Once I moved the app one folder up, it is
able to import the necessary modules.

A change to the subprocess call in PFT.py was needed to reflect this as
well.